src/pyptx/__init__.py
"""
pyptx – flex-style layout helper for python-pptx.
"""
from .errors import *
from .units import Inch, Centimeter, Ratio, Weight, resolve_length_span
from .core import Rect, LayoutItem, Box, Row, Column, Folds, SlideLayout

from pptx import Presentation
from pptx.dml.color import RGBColor
from pptx.enum.shapes import MSO_SHAPE
from pptx.util import Pt

__all__ = [
    "Inch", "Centimeter", "Ratio", "Weight", "resolve_length_span",
    "Rect", "LayoutItem", "Box", "Row", "Column", "Folds", "SlideLayout",
]
